# Speed up PackageClosureTests so not reported as long running