--- a/Core/Contributions/Refactory/Refactoring Browser/Environments/CategoryEnvironment.cls
+++ b/Core/Contributions/Refactory/Refactoring Browser/Environments/CategoryEnvironment.cls
@@ -17,7 +17,7 @@ categories
 	^categories select: [:each | self includesCategory: each]!
 
 categories: aCollection
-	categories := aCollection!
+	categories := aCollection asSet!
 
 classNames
 	^self categories inject: IdentitySet new