--- a/Core/Contributions/Refactory/Refactoring Browser/Environments/RBEnvironments.pax
+++ b/Core/Contributions/Refactory/Refactoring Browser/Environments/RBEnvironments.pax
@@ -151,7 +151,7 @@ isBehavior
 numberSelectors
 	"Answer the number of selectors in the receiver's method dictionary."
 
-	^(methodDictionary ?? #()) size!
+	^methodDictionary ifNil: [0] ifNotNil: [methodDictionary size]!
 
 selectorsAndMethodsDo: aDyadicValuable
 	"Evaluate the <dyadicValuable> argument for each of the selector/method pairs in