--- a/Core/DolphinVM/Compiler/stdafx.h
+++ b/Core/DolphinVM/Compiler/stdafx.h
@@ -55,3 +55,4 @@
 
 #include <vector>
 #include <iostream>
+#include <iomanip>