… Core/Contributions/Refactory/Refactoring Browser/Tests/BrowserEnvironmentTest.cls
… Core/Contributions/Refactory/Refactoring Browser/Tests/RBTests.pax
… Core/DolphinVM/Compiler/lexer.cpp
… Core/DolphinVM/Compiler/lexer.h
… Core/DolphinVM/Compiler/optimizer.cpp
… Core/DolphinVM/Compiler/str.h
… Core/DolphinVM/ObjMemPriv.inl
… Core/DolphinVM/STBehavior.h
… Core/DolphinVM/STClassDesc.h
… Core/DolphinVM/STObject.h
… Core/DolphinVM/STString.h
… Core/DolphinVM/SnapshotPrim.cpp
… Core/DolphinVM/StreamPrim.cpp
… Core/DolphinVM/StrgPrim.h
… Core/DolphinVM/TraceStream.h
… Core/DolphinVM/alloc.cpp
… Core/DolphinVM/byteasm.asm
… Core/DolphinVM/bytecde.cpp
… Core/DolphinVM/decode.cpp
… Core/DolphinVM/disassembler.h
… Core/DolphinVM/extcall.cpp
… Core/DolphinVM/objmem.h
… Core/DolphinVM/primitiv.cpp
… Core/DolphinVM/process.cpp
… Core/DolphinVM/realloc.cpp
… Core/DolphinVM/strgprim.cpp
… Core/Object Arts/Dolphin/ActiveX/Automation/ICreateErrorInfo.cls
… Core/Object Arts/Dolphin/ActiveX/Automation/IDispatch.cls
… Core/Object Arts/Dolphin/ActiveX/Automation/IEnumVARIANT.cls
… Core/Object Arts/Dolphin/ActiveX/Automation/IPerPropertyBrowsing.cls