Recently added PackageClosureTests>>testEditableListViewDemo was regularly taking 600ms. These changes in aggregate speed up these tests by around 30%.

Before:
```
00:04:08] PackageClosureTests>>testCommandLineHelloWorld ... PASSED in 143.6 ms
[00:04:08] PackageClosureTests>>testEditableListViewDemo ... PASSED in 602 ms
[00:04:09] PackageClosureTests>>testEtchASketch ... PASSED in 362.6 ms
[00:04:09] PackageClosureTests>>testHelloWorld ... PASSED in 276.3 ms
[00:04:09] PackageClosureTests>>testNotepad ... PASSED in 272 ms
[00:04:10] PackageClosureTests>>testRegEdit ... PASSED in 355.3 ms
[00:04:10] PackageClosureTests>>testSystemPackage ... PASSED in 79.8 ms
```

After:
```
[00:04:05] PackageClosureTests>>testCommandLineHelloWorld ... PASSED in 107.6 ms
[00:04:06] PackageClosureTests>>testEditableListViewDemo ... PASSED in 407.9 ms
[00:04:06] PackageClosureTests>>testEtchASketch ... PASSED in 272.6 ms
[00:04:06] PackageClosureTests>>testHelloWorld ... PASSED in 218.8 ms
[00:04:06] PackageClosureTests>>testNotepad ... PASSED in 214.5 ms
[00:04:07] PackageClosureTests>>testRegEdit ... PASSED in 266.7 ms
[00:04:07] PackageClosureTests>>testSystemPackage ... PASSED in 64.6 ms
```